+++ src/metadata/metadata_structs.h
@@ -390,41 +390,41 @@ struct ocf_metadata_iface {
 			ocf_cache_line_t line);
 
 	bool (*test_dirty)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop, bool all);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop, bool all);
 
 	bool (*test_out_dirty)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop);
 
 	bool (*clear_dirty)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop);
 
 	bool (*set_dirty)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop);
 
 	bool (*test_and_set_dirty)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop, bool all);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop, bool all);
 
 	bool (*test_and_clear_dirty)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop, bool all);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop, bool all);
 
 
 	bool (*test_valid)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop, bool all);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop, bool all);
 
 	bool (*test_out_valid)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop);
 
 	bool (*clear_valid)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop);
 
 	bool (*set_valid)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop);
 
 	bool (*test_and_set_valid)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop, bool all);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop, bool all);
 
 	bool (*test_and_clear_valid)(struct ocf_cache *cache,
-		ocf_cache_line_t line, uint8_t start, uint8_t stop, bool all);
+		ocf_cache_line_t line, uint32_t start, uint32_t stop, bool all);
 };
 
 struct ocf_cache_line_settings {
